refactor: moving controller
